pyproject.toml
[tool.poetry]
name = "hdforce"
version = "0.0.1"
description = "Get your data from the Hawkin Dynamics API"
authors = ["laureng-hd <lauren@hawkindynamics.com>"]
readme = "README.md"
keywords = ["hawkin", "force plates", "sports science"]
homepage = "https://www.hawkindynamics.com/"
repository = "https://github.com/HawkinDynamics/hawkinPy"
documentation = "https://hawkindynamics.github.io/hawkinPy/index.html"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]


[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
datetime = "^5.5"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
